+++ unitlinks/src/apdu_frame.c
@@ -18,7 +18,6 @@
 
 apdu_frame *apdu_frame_create()
 {
-	// try to allocate memory for the structure
 	apdu_frame *frame = (apdu_frame*) calloc(1, sizeof(apdu_frame));
 
 	return frame;
@@ -27,10 +26,8 @@ apdu_frame *apdu_frame_create()
 
 void apdu_frame_cleanup(apdu_frame *frame)
 {
-	// fast check input data
 	if(!frame || !frame->data) return;
 
-	// free memory allocated for the data
 	frame->data_len = 0;
 	free(frame->data);
 	frame->data = NULL;
@@ -39,10 +36,8 @@ void apdu_frame_cleanup(apdu_frame *frame)
 
 void apdu_frame_destroy(apdu_frame **frame)
 {
-	// fast check input data
 	if(!*frame) return;
 
-	// cleanup insides
 	apdu_frame_cleanup(*frame);
 
 	free(*frame);
@@ -61,25 +56,20 @@ void apdu_frame_buff_parse_ctrl_field(unsigned char *buff, uint32_t *offset, apd
 	switch(frame->type)
 	{
 	case APCI_TYPE_I:
-		// get send sequence number
 		frame->send_num = buff_get_le_uint16(buff, *offset) >> 1;
 
-		// get receive sequence number
 		frame->recv_num = buff_get_le_uint16(buff, *offset + 2) >> 1;
 		break;
 
 	case APCI_TYPE_S:
-		// get receive sequence number
 		frame->recv_num = buff_get_le_uint16(buff, *offset + 2) >> 1;
 		break;
 
 	case APCI_TYPE_U:
-		// get U-Format command code
 		frame->u_cmd    = buff_get_le_uint8(buff, *offset) & ~0x03;
 		break;
 	}
 
-	// move over 4 bytes of control fields
 	*offset += 4;
 }
 
@@ -89,23 +79,18 @@ void apdu_frame_buff_build_ctrl_field(unsigned char *buff, uint32_t *offset, apd
 	switch(frame->type)
 	{
 	case APCI_TYPE_I:
-		// put send sequence number & frame type
 		buff_put_le_uint16(buff, *offset, (frame->send_num << 1) | frame->type);
 
-		// put receive sequence number
 		buff_put_le_uint16(buff, *offset + 2, frame->recv_num << 1);
 		break;
 
 	case APCI_TYPE_S:
-		// put frame type
 		buff_put_le_uint16(buff, *offset, frame->type);
 
-		// put receive sequence number
 		buff_put_le_uint16(buff, *offset + 2, frame->recv_num << 1);
 		break;
 
 	case APCI_TYPE_U:
-		// put U-Format command code
 		buff_put_le_uint8(buff, *offset, frame->u_cmd | frame->type);
 
 		// set other 3 bytes to zero
@@ -116,17 +101,16 @@ void apdu_frame_buff_build_ctrl_field(unsigned char *buff, uint32_t *offset, apd
 		break;
 	}
 
-	// move over 4 bytes control field
 	*offset += 4;
 }
 
 
-uint8_t apdu_frame_buff_parse(unsigned char *buff, uint32_t buff_len, uint32_t *offset, apdu_frame *frame)
+uint16_t apdu_frame_buff_parse(unsigned char *buff, uint32_t buff_len, uint32_t *offset, apdu_frame *frame)
 {
 	// fast check input data
 	if(!buff || !frame) return RES_INCORRECT;
 
-	uint8_t res;
+	uint16_t res;
 	uint8_t start_byte = 0;
 
 	// look for the frame start
@@ -134,7 +118,6 @@ uint8_t apdu_frame_buff_parse(unsigned char *buff, uint32_t buff_len, uint32_t *
 	{
 		start_byte = buff_get_le_uint8(buff, *offset);
 
-		// move to the next byte
 		*offset += 1;
 
 		if(start_byte == APDU_START_BYTE) break;
@@ -143,10 +126,8 @@ uint8_t apdu_frame_buff_parse(unsigned char *buff, uint32_t buff_len, uint32_t *
 	// check if rest of the buffer long enough to contain the frame
 	if(buff_len - *offset + 1 < APDU_LEN_MIN) return RES_LEN_INVALID;
 
-	// check if start byte found, otherwise return error
 	if(start_byte != APDU_START_BYTE) return RES_INCORRECT;
 
-	// get data length
 	frame->data_len = buff_get_le_uint8(buff, *offset) - 4;
 	*offset += 1;
 
@@ -158,22 +139,17 @@ uint8_t apdu_frame_buff_parse(unsigned char *buff, uint32_t buff_len, uint32_t *
 		// check if data length is correct
 		if(frame->data_len > 0 && buff_len - *offset >= frame->data_len)
 		{
-			// allocate memory for the data
 			frame->data = (unsigned char*) malloc(frame->data_len);
 
-			// check if memory allocated OK, otherwise return error
 			if(!frame->data)
 			{
-				// set data length to zero
 				frame->data_len = 0;
 
 				return RES_MEM_ALLOC;
 			}
 
-			// copy data from the buffer
 			memcpy((void*)frame->data, (void*)(buff + *offset), frame->data_len);
 
-			// move over data in the buffer
 			*offset += frame->data_len;
 
 			res = RES_SUCCESS;
@@ -188,14 +164,12 @@ uint8_t apdu_frame_buff_parse(unsigned char *buff, uint32_t buff_len, uint32_t *
 			// set offset to the end of the buffer
 			*offset = buff_len;
 
-			// return error
 			res = RES_LEN_INVALID;
 		}
 		break;
 
 	case APCI_TYPE_S:
 	case APCI_TYPE_U:
-		// no data with this type of frame therefore set data pointer and length to zero
 		frame->data_len = 0;
 		frame->data = NULL;
 
@@ -211,25 +185,19 @@ uint8_t apdu_frame_buff_parse(unsigned char *buff, uint32_t buff_len, uint32_t *
 }
 
 
-uint8_t apdu_frame_buff_build(unsigned char **buff, uint32_t *buff_len, apdu_frame *frame)
+uint16_t apdu_frame_buff_build(unsigned char **buff, uint32_t *buff_len, apdu_frame *frame)
 {
-	// set buffer length to zero and start building it
 	*buff_len = 0;
 	uint32_t offset = 0;
 
-	// fast check input data
 	if(!buff || !frame || (frame->type == APCI_TYPE_I && (frame->data_len == 0 || frame->data == NULL))) return RES_INCORRECT;
 
-	// calculate buffer length
 	*buff_len = 6 + frame->data_len;
 
-	// allocate memory for the buffer
 	*buff = (unsigned char*) malloc(*buff_len);
 
-	// check if memory allocated OK, otherwise return error
 	if(!*buff)
 	{
-		// set buffer length to zero
 		*buff_len = 0;
 
 		return RES_MEM_ALLOC;
@@ -240,7 +208,6 @@ uint8_t apdu_frame_buff_build(unsigned char **buff, uint32_t *buff_len, apdu_fra
 	buff_put_le_uint8(*buff, offset, APDU_START_BYTE);
 	offset += 1;
 
-	// put data length in the buffer
 	buff_put_le_uint8(*buff, offset, *buff_len - 2);
 	offset += 1;
 
@@ -248,7 +215,6 @@ uint8_t apdu_frame_buff_build(unsigned char **buff, uint32_t *buff_len, apdu_fra
 
 	if(frame->type == APCI_TYPE_I)
 	{
-		// copy data to buffer
 		memcpy((void*)(*buff + offset), (void*)frame->data, frame->data_len);
 		offset += frame->data_len;
 	}
